+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix border radius clipping issue
+- Image component accessibility issue when alt value is empty
 
 ## [0.4.0] - 2025-02-27